requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
umap-learn>=0.5.4
fast-hdbscan>=0.1.3
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.10.0
tqdm>=4.65.0
joblib>=1.3.0
